Department Service Update Include

+++ SerenityHospital.Business/Services/Implements/DepartmentService.cs
@@ -4,6 +4,7 @@ using SerenityHospital.Business.Dtos.DepartmentDtos;
 using SerenityHospital.Business.Exceptions.Common;
 using SerenityHospital.Business.Exceptions.Departments;
 using SerenityHospital.Business.Exceptions.Images;
+using SerenityHospital.Business.Exceptions.PatientRooms;
 using SerenityHospital.Business.Extensions;
 using SerenityHospital.Business.ExternalServices.Interfaces;
 using SerenityHospital.Business.Services.Interfaces;
@@ -15,16 +16,18 @@ namespace SerenityHospital.Business.Services.Implements;
 public class DepartmentService : IDepartmentService
 {
     readonly IDepartmentRepository _repo;
+    readonly IPatientRoomRepository _patientRepo;
     readonly IMapper _mapper;
     readonly IFileService _fileService;
     readonly IServiceRepository _serviceRepo;
 
-    public DepartmentService(IDepartmentRepository repo, IMapper mapper, IFileService fileService, IServiceRepository serviceRepo)
+    public DepartmentService(IDepartmentRepository repo, IMapper mapper, IFileService fileService, IServiceRepository serviceRepo, IPatientRoomRepository patientRepo)
     {
         _repo = repo;
         _mapper = mapper;
         _fileService = fileService;
         _serviceRepo = serviceRepo;
+        _patientRepo = patientRepo;
     }
 
     public async Task CreateAsync(DepartmentCreateDto dto)
@@ -46,8 +49,10 @@ public class DepartmentService : IDepartmentService
     public async Task DeleteAsync(int id)
     {
         if (id <= 0) throw new NegativeIdException<Department>();
-        var entity = await _repo.GetByIdAsync(id);
+        var entity = await _repo.GetByIdAsync(id,"PatientRooms");
         if (entity is null) throw new NotFoundException<Department>();
+
+        if (entity.PatientRooms.Count() > 0) throw new DepartmentIsNotEmptyException();
 
         _repo.Delete(entity);
         _fileService.Delete(entity.IconUrl);
@@ -58,11 +63,11 @@ public class DepartmentService : IDepartmentService
     {
         if(takeAll)
         {
-            return _mapper.Map<IEnumerable<DepartmentListItemDto>>(_repo.GetAll());
+            return _mapper.Map<IEnumerable<DepartmentListItemDto>>(_repo.GetAll("PatientRooms"));
         }
         else
         {
-            return _mapper.Map<IEnumerable<DepartmentListItemDto>>(_repo.FindAll(d => d.IsDeleted == false));
+            return _mapper.Map<IEnumerable<DepartmentListItemDto>>(_repo.FindAll(d => d.IsDeleted == false,"PatientRooms"));
         }
     }
 
@@ -72,12 +77,12 @@ public class DepartmentService : IDepartmentService
         Department? entity;
         if(takeAll)
         {
-            entity = await _repo.GetByIdAsync(id);
+            entity = await _repo.GetByIdAsync(id,"PatientRooms");
             if (entity is null) throw new NotFoundException<Department>();
         }
         else
         {
-            entity = await _repo.GetSingleAsync(d => d.Id == id && d.IsDeleted == false);
+            entity = await _repo.GetSingleAsync(d => d.Id == id && d.IsDeleted == false, "PatientRooms");
             if (entity is null) throw new NotFoundException<Department>();
         }
 
@@ -87,8 +92,10 @@ public class DepartmentService : IDepartmentService
     public async Task ReverteSoftDeleteAsync(int id)
     {
         if (id <= 0) throw new NegativeIdException<Department>();
-        var entity = await _repo.GetByIdAsync(id);
+        var entity = await _repo.GetByIdAsync(id,"PatientRooms");
         if (entity is null) throw new NotFoundException<Department>();
+
+        if (entity.PatientRooms.Count() > 0) throw new DepartmentIsNotEmptyException();
 
         _repo.RevertSoftDelete(entity);
         await _repo.SaveAsync();
@@ -97,8 +104,10 @@ public class DepartmentService : IDepartmentService
     public async Task SoftDeleteAsync(int id)
     {
         if (id <= 0) throw new NegativeIdException<Department>();
-        var entity = await _repo.GetByIdAsync(id);
+        var entity = await _repo.GetByIdAsync(id,"PatientRooms");
         if (entity is null) throw new NotFoundException<Department>();
+
+        if (entity.PatientRooms.Count() > 0) throw new DepartmentIsNotEmptyException();
 
         _repo.SoftDelete(entity);
         await _repo.SaveAsync();
@@ -117,6 +126,15 @@ public class DepartmentService : IDepartmentService
             if (!dto.IconFile.IsSizeValid(3)) throw new SizeNotValidException();
             if (!dto.IconFile.IsTypeValid("image")) throw new TypeNotValidException();
             entity.IconUrl = await _fileService.UploadAsync(dto.IconFile, RootConstant.DepartmentImageRoot);
+        }
+
+        foreach (var itemId in dto.PatientRoomIds)
+        {
+            var patientRoom =await _patientRepo.GetByIdAsync(itemId);
+            if (patientRoom is null) throw new NotFoundException<PatientRoom>();
+            var isPatientRoomInOtherDepartment = await _repo.IsExistAsync(d => d.Id != id && d.PatientRooms.Any(d => d.Id == patientRoom.Id));
+            if (isPatientRoomInOtherDepartment) throw new PatientRoomInOtherDepartmentException();
+            entity?.PatientRooms.Add(patientRoom);
         }
 
         var service = await _serviceRepo.GetByIdAsync(dto.ServiceId);
